Warning int vs. size_t
git-svn-id: svn://svn.berlios.de/lincity-ng/trunk@1707 8226ee30-1aee-0310-a633-b68f58d1eb45

+++ src/gui/Paragraph.cpp
@@ -539,7 +539,7 @@ Paragraph::setText(const std::string& newtext, const Style& style)
         delete *i;
     textspans.clear();
 
-    int span_end = newtext.find_first_of('\t',0);
+    size_t span_end = newtext.find_first_of('\t',0);
     if(span_end != newtext.npos) // we have a tab
     {
         int tabcount = 0;
